requirements.txt
streamlit
pandas
matplotlib
python-bidi
google-genai
openpyxl